new-pkg:(formatter) move code to a new go pkg
